+++ src/modules/clients/controllers/admin/contacts.controller.ts
@@ -7,43 +7,66 @@ import {
   Param,
   Delete,
   ValidationPipe,
+  NotFoundException,
 } from '@nestjs/common';
 import { ContactsService } from '../../services/contacts.service';
 import { CreateContactDto } from '../../dto/create-contact.dto';
 import { UpdateContactDto } from '../../dto/update-contact.dto';
 import { IdPipe } from '../../../core/transformers/id.pipe';
 
-@Controller('admin/contacts')
+@Controller('admin/clients/:clientId/contacts')
 export class ContactsController {
   constructor(private readonly contactsService: ContactsService) {}
 
   @Post()
-  create(@Body(ValidationPipe) createContactDto: CreateContactDto) {
-    return this.contactsService.create(createContactDto);
+  create(
+    @Param('clientId') clientId: string,
+    @Body(ValidationPipe) createContactDto: CreateContactDto,
+  ) {
+    return this.contactsService.create(+clientId, createContactDto);
   }
 
   @Get()
-  findAll() {
-    return this.contactsService.findAll();
+  findAll(@Param('clientId') clientId: string) {
+    return this.contactsService.findAll(+clientId);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.contactsService.findOne(+id);
+  async findOne(@Param('clientId') clientId: string, @Param('id') id: string) {
+    const contact = await this.contactsService.findOne(+clientId, +id);
+
+    if (!contact) {
+      throw new NotFoundException('Contact not found');
+    }
+
+    return contact;
   }
 
   @Patch(':id')
   async update(
+    @Param('clientId') clientId: string,
     @Param('id') id: string,
     @Body(IdPipe, ValidationPipe) updateContactDto: UpdateContactDto,
   ) {
+    const contact = await this.contactsService.findOne(+clientId, +id);
+
+    if (!contact) {
+      throw new NotFoundException('Contact not found');
+    }
+
     await this.contactsService.update(+id, updateContactDto);
 
-    return this.contactsService.findOne(+id);
+    return this.contactsService.findOne(+clientId, +id);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('clientId') clientId: string, @Param('id') id: string) {
+    const contact = await this.contactsService.findOne(+clientId, +id);
+
+    if (!contact) {
+      throw new NotFoundException('Contact not found');
+    }
+
     return this.contactsService.remove(+id);
   }
 }
